Fix deliver stage (#2)
A bug in Docusaurus causes it to only look at the global git config. Set the name and email globally, but only if it is running in CI by checking the $CI variable for "true". Codefresh sets this variable automatically, and it is compatible with any other CI engine as well.